Refactor: Move visitors to ast
